pyproject.toml
[tool.poetry]
name = "eduplatform"
version = "0.1.0"
description = ""
authors = ["Marina Kondratyk <kondratsiuk.marina@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.3"
pillow = "^10.0.0"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.7.0"
flake8 = "^6.0.0"
pyupgrade = "^3.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
